Updated Business Logic

+++ src/main/java/example/demo/springboot/helloworld/controller/HomeController.java
@@ -14,9 +14,9 @@ public class HomeController {
         return "Hello from Test CI/CD!";
     }
 
-    private uintmax_t accumulate(uintmax_t max) {
-        uintmax_t ret;
-        for(uintmax_t i = 1; i <= max; i++) {
+    private long accumulate(int) max) {
+        long ret;
+        for(long i = 1; i <= max; i++) {
             ret += i;
         }
         return ret;
